Reorganize server to move design file related functions into dedicated folder.
